rename : InputHandler를 VoucherValidator로 이름 변경
